refactor: reorganize GraphQL DTOs and update imports for consistency
