+++ src/graphql/dtos/wallet-balance.dto.ts
